# apps/vuln-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title:  TestBed API (api-demo)
  version: 1.0.0
  description: >-
    Static OpenAPI spec for the api-demo app. For full narrative docs see
    API-DOCUMENTATION.md. Endpoints are intentionally vulnerable for testing.
servers:
  - url: http://localhost:8880
...
paths:
...
  /api/v1/telecom/billing/adjustments:
    put:
      tags: [Telecom]
      summary: Billing adjustment tampering
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/telecom/billing/invoices/{id}:
    get:
      tags: [Telecom]
      summary: Invoice IDOR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/telecom/billing/payment-methods:
    post:
      tags: [Telecom]
      summary: Payment method bypass
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/telecom/billing/refunds:
    post:
      tags: [Telecom]
      summary: Refund abuse
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/telecom/devices/bind:
    post:
      tags: [Telecom]
      summary: Device binding bypass
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/telecom/imei/blacklist:
    put:
      tags: [Telecom]
      summary: IMEI blacklist override
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/telecom/integrations/cdr/stream:
    post:
      tags: [Telecom]
      summary: CDR stream bypass
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/telecom/integrations/device-activate:
    post:
      tags: [Telecom]
      summary: Device activation bypass
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/telecom/integrations/webhooks:
    post:
      tags: [Telecom]
      summary: "Webhook registration (SSRF risk)"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/v1/telecom/network/provision:
    post:
      tags: [Telecom]
      summary: Provisioning bypass
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/telecom/network/throttle:
    put:
      tags: [Telecom]
      summary: Throttle manipulation
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/telecom/network/towers/{id}:
    get:
      tags: [Telecom]
      summary: "Tower access (IDOR)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/telecom/porting/requests:
    post:
      tags: [Telecom]
      summary: Port-out bypass
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/telecom/porting/requests/{id}:
    get:
      tags: [Telecom]
      summary: "Porting status (IDOR)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/telecom/porting/swap:
    put:
      tags: [Telecom]
      summary: Number swap tampering
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/telecom/roaming/override:
    post:
      tags: [Telecom]
      summary: Roaming override
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/v1/telecom/subscribers/{id}/plan:
    put:
      tags: [Telecom]
      summary: Plan change tampering
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/telecom/subscribers/{id}/profile:
    get:
      tags: [Telecom]
      summary: "Subscriber profile (IDOR)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/telecom/subscribers/{id}/sim-swap:
    post:
      tags: [Telecom]
      summary: SIM swap bypass
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...
  /api/v1/telecom/billing/invoices/{invoice_id}:
    get:
      tags: [Telecom]
      summary: Get
      parameters:
        - name: invoice_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/telecom/network/towers/{tower_id}:
    get:
      tags: [Telecom]
      summary: Get
      parameters:
        - name: tower_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/telecom/porting/requests/{request_id}:
    get:
      tags: [Telecom]
      summary: Get
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/telecom/subscribers/{subscriber_id}/plan:
    put:
      tags: [Telecom]
      summary: Put
      parameters:
        - name: subscriber_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
  /api/v1/telecom/subscribers/{subscriber_id}/profile:
    get:
      tags: [Telecom]
      summary: Get
      parameters:
        - name: subscriber_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/telecom/subscribers/{subscriber_id}/sim-swap:
    post:
      tags: [Telecom]
      summary: Post
      parameters:
        - name: subscriber_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...